+++ tests/qiskit_convert_test.py
@@ -35,7 +35,6 @@ from qiskit.transpiler.passes import BasisTranslator  # type: ignore
 from qiskit.circuit.equivalence_library import StandardEquivalenceLibrary  # type: ignore
 from qiskit_ibm_runtime.fake_provider import FakeGuadalupe  # type: ignore
 from qiskit.circuit.parameterexpression import ParameterExpression  # type: ignore
-from pytket.extensions.qiskit import qiskit_to_tk, tk_to_qiskit
 from qiskit.circuit.library import TwoLocal
 from qiskit import transpile
 
@@ -53,6 +52,7 @@ from pytket.circuit import (
     StatePreparationBox,
 )
 from pytket.extensions.qiskit import tk_to_qiskit, qiskit_to_tk, IBMQBackend
+from pytket.extensions.qiskit.backends import qiskit_aer_backend
 from pytket.extensions.qiskit.qiskit_convert import _gate_str_2_optype
 from pytket.extensions.qiskit.tket_pass import TketPass, TketAutoPass
 from pytket.extensions.qiskit.result_convert import qiskit_result_to_backendresult
@@ -238,7 +238,7 @@ def test_symbolic() -> None:
 
 def test_measures() -> None:
     qc = get_test_circuit(True)
-    backend = Aer.get_backend("aer_simulator")
+    backend = qiskit_aer_backend("aer_simulator")
     qc = transpile(qc, backend)
     job = backend.run([qc], seed_simulator=7)
     counts0 = job.result().get_counts(qc)
@@ -374,7 +374,7 @@ def test_tketpass() -> None:
 def test_tketautopass(brisbane_backend: IBMQBackend) -> None:
     backends = [
         Aer.get_backend("aer_simulator_statevector"),
-        Aer.get_backend("aer_simulator"),
+        qiskit_aer_backend("aer_simulator"),
         Aer.get_backend("aer_simulator_unitary"),
     ]
     backends.append(brisbane_backend._backend)
@@ -559,7 +559,7 @@ def test_convert_result() -> None:
     qc.measure(qr1[0], cr[0])
     qc.measure(qr2[1], cr2[0])
 
-    simulator = Aer.get_backend("aer_simulator")
+    simulator = qiskit_aer_backend("aer_simulator")
     qisk_result = simulator.run(qc, shots=10).result()
 
     tk_res = next(qiskit_result_to_backendresult(qisk_result))